+++ packages/dataviews/src/field-types/media.tsx
@@ -1,20 +1,55 @@
 /**
  * Internal dependencies
  */
-import type { FieldTypeDefinition } from '../types';
+import type { Field, NormalizedField, Rules } from '../types';
+import { getControl } from '../dataform-controls';
+import hasElements from './utils/has-elements';
+import getValueFromId from './utils/get-value-from-id';
+import setValueFromId from './utils/set-value-from-id';
 
 function sort() {
 	return 0;
 }
 
-export default {
-	sort,
-	isValid: {
+function render() {
+	return null;
+}
+
+export default function normalizeField< Item >(
+	field: Field< Item >
+): NormalizedField< Item > {
+	const getValue = field.getValue || getValueFromId( field.id );
+	const setValue = field.setValue || setValueFromId( field.id );
+	const isValid: Rules< Item > = {
 		elements: true,
 		custom: () => null,
-	},
-	Edit: null,
-	render: () => null,
-	enableSorting: false,
-	filterBy: false,
-} satisfies FieldTypeDefinition< any >;
+	};
+
+	return {
+		id: field.id,
+		type: 'media',
+		label: field.label || field.id,
+		header: field.header || field.label || field.id,
+		description: field.description,
+		placeholder: field.placeholder,
+		getValue,
+		setValue,
+		elements: field.elements,
+		getElements: field.getElements,
+		hasElements: hasElements( field ),
+		render: field.render ?? render,
+		Edit: getControl( field, null ),
+		sort: field.sort ?? sort,
+		isValid: {
+			...isValid,
+			...field.isValid,
+		},
+		isVisible: field.isVisible,
+		enableSorting: field.enableSorting ?? false,
+		enableGlobalSearch: field.enableGlobalSearch ?? false,
+		enableHiding: field.enableHiding ?? true,
+		readOnly: field.readOnly ?? false,
+		filterBy: false,
+		format: {},
+	};
+}